+++ docs/source/release_notes.rst
@@ -33,6 +33,9 @@ Upcoming Release
   - **Update notebooks to always install latest repo version from Google Colab**: This ensures that the latest version is always used
 
 
+  **New Features:**
+  - **Generic Forecasting Interface**: This interface enables to specify different forecast algorithms for preprocess, initialization and update during runtime. They can be specified in the config.yaml or unit csv files. For more information about currently implemented algorithms and how to specify them please read the documentation on Unit forecasts.
+
 0.5.6 - (23th December 2025)
 ============================